Novo/src/app/backEnd/requirements.txt
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
pickle-mixin==1.0.2
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
Werkzeug==2.2.2
kafka-python==2.0.2

